requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
